fix recursive import in test
