+++ pyworkflow/em/packages/chimera/protocol_fit.py
@@ -26,23 +26,8 @@
 # *
 # **************************************************************************
 
-import os
-from pyworkflow import VERSION_1_2
-from pyworkflow.em import PdbFile
-from pyworkflow.em import Volume
-from pyworkflow.em.convert import ImageHandler
-from pyworkflow.em.protocol import EMProtocol
-from pyworkflow.em.convert_header.CCP4.convert import Ccp4Header
-from pyworkflow.em.viewers.chimera_utils import \
-    createCoordinateAxisFile, \
-    adaptOriginFromCCP4ToChimera, getProgram, runChimeraProgram,\
-    chimeraPdbTemplateFileName, chimeraMapTemplateFileName, \
-    chimeraScriptFileName
-from pyworkflow.em.packages.chimera.protocol_operate import createScriptFile
-from pyworkflow.protocol.params import MultiPointerParam, PointerParam, \
-    StringParam
-from pyworkflow.utils.properties import Message
-from pyworkflow.em.packages.chimera.protocol_operate import ChimeraProtOperate
+from protocol_operate import ChimeraProtOperate
+#from pyworkflow.em.packages.chimera.protocol_operate import ChimeraProtOperate
 
 
 class ChimeraProtRigidFit(ChimeraProtOperate):
